Added packages for models, controllers and services
