+++ .pre-commit-config.yaml
@@ -14,3 +14,9 @@ repos:
     rev: 6.1.1  # pick a git hash / tag to point to
     hooks:
     -   id: pydocstyle
+-   repo: https://github.com/pycqa/isort
+    rev: 5.11.2
+    hooks:
+    -   id: isort
+        args: ["--profile", "black"]
+        name: isort (python)
