+++ backend/pyproject.toml
@@ -24,6 +24,7 @@ pre-commit = "^2.20.0"
 pydocstyle = "^6.1.1"
 toml = "^0.10.2"
 faker = "^15.3.4"
+isort = "^5.11.3"
 
 [tool.poe.tasks]
 test = "pytest"
