+++ backend/pyrightconfig.json
@@ -1,0 +1,4 @@
+{
+    "venvPath": "/home/pspiagicw/.cache/pypoetry/virtualenvs",
+    "venv": "placy-toW1y4vv-py3.10",
+}
